Adjust buildkite link regex
Currently we require the `#UUID` fragment when matching buildkite links
but the issues created by this tool don't contain them, so they will
be treated as new issues even when filed.

Just remove the fragment part of the regex for now, as we only used the
pre-fragment part of the regex anyway for further processing.